deep_pysearch/requirements.txt
aiohttp>=3.8.5
python-dotenv>=1.0.0
